Move verifier from internal and grpc to subpackage
We now export verifier, but make no stability guarantees.
Additionally, we move grpc_client.go and other grpc-related files to the
grpc/ subpackage to better couple all verifier implementation details
with the protocol.
